+++ cantera18/branches/scons/ext/f2c_libs/lwrite.c
@@ -169,7 +169,7 @@ l_g(char *buf, double n)
 			}
 		}
  f__ret:
-	return b - buf;
+	return (int) (b - buf);
 #endif
 	}
